Append origin server to display account name
This PR appends the name of the origin server (taken from the StorageURL configuration item) to all display account names. This shows up in Xochitl’s general settings panel. This is useful to know at a glance whether you’re connected to the official servers or an rmfakecloud server, and if so, to which rmfakecloud server.

Tested on my personal rmfakecloud instance with a connected rM2 running system version 2.10.1.332.